+++ view/frontend/web/js/view/payment/gateway/method-renderer/einvoicing.js
@@ -18,6 +18,8 @@ define(
     [
         'jquery',
         'Magento_Checkout/js/view/payment/default',
+        'Magento_Checkout/js/action/select-payment-method',
+        'Magento_Checkout/js/checkout-data',
         'Magento_Checkout/js/action/redirect-on-success',
         'mage/url'
     ],
@@ -26,6 +28,8 @@ define(
      *
      * @param $
      * @param Component
+     * @param selectPaymentMethodAction
+     * @param checkoutData
      * @param redirectOnSuccessAction
      * @param url
      * @returns {*}
@@ -33,6 +37,8 @@ define(
     function (
         $,
         Component,
+        selectPaymentMethodAction,
+        checkoutData,
         redirectOnSuccessAction,
         url
     ) {
@@ -53,6 +59,12 @@ define(
                 this._super()
                     .observe('dateOfBirth')
                     .observe('accountNumber');
+
+                if (config.is_preselected) {
+                    selectPaymentMethodAction(this.getData());
+                    checkoutData.setSelectedPaymentMethod(this.item.method);
+                }
+
                 self = this;
             },